requirements.txt
asttokens==3.0.1
audioread==3.0.1
backcall==0.2.0
certifi==2026.2.25
cffi==1.17.1
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
filelock==3.16.1
fsspec==2025.3.0
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.8.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.41.1
matplotlib-inline==0.1.7
msgpack==1.1.1
nest-asyncio==1.6.0
numba==0.58.1
numpy==1.24.4
opencv-python==4.9.0.80
packaging==26.0
pandas==2.0.3
parso==0.8.6
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
soundfile==0.13.1
soxr==0.3.7
stack-data==0.6.3
threadpoolctl==3.5.0
tokenizers==0.15.2
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.4.2
tqdm==4.67.3
traitlets==5.14.3
transformers==4.36.2
typing_extensions==4.13.2
tzdata==2025.3
urllib3==2.2.3
wcwidth==0.6.0
zipp==3.20.2